Made index.js the entrypoint
