+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.3.2",
   "type": "module",
   "description": "Apps Script IDE framework",
-  "bin": "dist/src/app.js",
+  "bin": "dist/src/index.js",
   "files": [
     "dist/src",
     "dist/.gitignore-target",
